Fixed crash due to bad multi threading locks in saving mcf files

+++ src/shared/mcfcore/code/thread/BaseMCFThread.h
@@ -101,7 +101,7 @@ namespace MCFCore
 			std::mutex m_pFileMutex;
 
 			std::vector<std::shared_ptr<MCFCore::MCFFile>> &m_rvFileList;
-			std::vector<int32> m_vFileList;
+			std::vector<size_t> m_vFileList;
 
 			uint64 m_uiFileOffset;
 		};
